后台权限完善

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -6,19 +6,21 @@ use Spatie\Permission\Models\Permission;
 
 class PermissionController extends BaseController
 {
+    public function __construct()
+    {
+        parent::__construct();
+        $this->middleware('managePermissions');
+    }
     public function index(){
-//        $this->authorize('permission_index');
         $permissions=Permission::paginate(10);
         return view('admin.permission.index', compact('permissions'));
     }
 
     public function create(){
-//        $this->authorize('permission_add');
         return view('admin.permission.create');
     }
 
     public function store(Request $request, Permission $permission){
-//        $this->authorize('permission_add');
         $permission->name=$request->name;
         $permission->title=$request->title;
         $permission->save();
@@ -26,12 +28,10 @@ class PermissionController extends BaseController
     }
 
     public function edit(Permission $permission){
-//        $this->authorize('permission_edit');
         return view('admin.permission.create', compact('permission'));
     }
 
     public function update(Permission $permission, Request $request){
-//        $this->authorize('permission_edit');
         $permission->name=$request->name;
         $permission->title=$request->title;
         $permission->save();
@@ -39,7 +39,7 @@ class PermissionController extends BaseController
     }
 
     public function destroy(Permission $permission){
-//        $this->authorize('permission_del');
+        $this->authorize('del_permission');
         $permission->delete();
         return back()->with('success','删除权限成功');
     }
